% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_complete}
\alias{check_complete}
\title{Check Complete}
\usage{
check_complete(x, floored = TRUE, unique = TRUE, sorted = TRUE,
  units = dtt_units(x), x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the values must be sorted.}

\item{units}{A string of the units.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks whether an object x is complete using \code{dtt_completed()}.
}
\examples{
check_complete(Sys.Date())
}
\seealso{
\code{\link{dtt_completed}}
}
