% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dtt}
\alias{check_dtt}
\title{Check DateTime Vector}
\usage{
check_dtt(x, values = NULL, length = NA, unique = FALSE,
  sorted = FALSE, floored = FALSE, complete = FALSE,
  tz = dtt_tz(x), units = dtt_units(x), named = NA,
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{values}{NULL or a vector specifying the values.}

\item{length}{A flag indicating whether x should have elements (versus no elements) or a missing value indicating no requirements or a count or count range of the number of elements or a count vector of the permitted number of elements.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{floored}{A flag indicating whether x should be floored.}

\item{complete}{A flag indicating whether x should be complete.}

\item{tz}{A string of the time zone.}

\item{units}{A string of the units.}

\item{named}{A flag indicating whether the vector must be named or unnamed or NA if it doesn't matter.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks whether object x is a Date or POSIXct vector.
}
\examples{
check_dtt(Sys.Date())
check_dtt(Sys.time())
}
\seealso{
\code{\link[checkr]{check_vector}}
}
