% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_floored}
\alias{check_floored}
\title{Check Floored}
\usage{
check_floored(x, units, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{units}{A string of the time units to floor by.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks whether an object x is floored using \code{dtt_floored()}.
}
\examples{
check_floored(Sys.time(), "days", error = FALSE)
}
\seealso{
\code{\link{dtt_floored}}
}
