% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_rate}
\alias{check_rate}
\title{Check Rate}
\usage{
check_rate(x, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks an object is a \code{dtt_rate}.
}
\examples{
check_rate(dtt_rate(1.5))
}
\seealso{
\code{\link{dtt_rate}}
}
