% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_time_units}
\alias{check_time_units}
\title{Check Time Units}
\usage{
check_time_units(x, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks an object is string of the possible time units.
}
\examples{
check_time_units("hours")
}
