% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_timer}
\alias{check_timer}
\title{Check Timer}
\usage{
check_timer(x, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks an object is a \code{dtt_timer}.
}
\examples{
check_timer(dtt_timer(1L))
}
\seealso{
\code{\link{dtt_duration}}
}
