% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_times}
\alias{date_times}
\title{Date Times}
\format{An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 3.}
\usage{
date_times
}
\description{
A POSIXct vector of length 3.
}
\keyword{datasets}
