% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{dtt_add_days}
\alias{dtt_add_days}
\title{Add Days}
\usage{
dtt_add_days(x, n = 1L, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{n}{An integer of the number of days to add}

\item{...}{Unused.}
}
\value{
The modified Date or POSIXct vector.
}
\description{
Add Days
}
\examples{
dtt_add_days(Sys.time())
}
