% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{dtt_add_hours}
\alias{dtt_add_hours}
\title{Add Hours}
\usage{
dtt_add_hours(x, n = 1L, ...)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{n}{An integer of the number of hours to add}

\item{...}{Unused.}
}
\value{
The modified POSIXct vector.
}
\description{
Add Hours
}
\examples{
dtt_add_hours(Sys.time())
}
