% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{dtt_add_minutes}
\alias{dtt_add_minutes}
\title{Add Minutes}
\usage{
dtt_add_minutes(x, n = 1L, ...)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{n}{An integer of the number of minutes to add}

\item{...}{Unused.}
}
\value{
The modified POSIXct vector.
}
\description{
Add Minutes
}
\examples{
dtt_add_minutes(Sys.time())
}
