% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{dtt_add_seconds}
\alias{dtt_add_seconds}
\title{Add Seconds}
\usage{
dtt_add_seconds(x, n = 1L, ...)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{n}{An integer of the number of seconds to add}

\item{...}{Unused.}
}
\value{
The modified POSIXct vector.
}
\description{
Add Seconds
}
\examples{
dtt_add_seconds(Sys.time())
}
