% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/subtract.R
\name{dtt_add_units}
\alias{dtt_add_units}
\alias{dtt_subtract_units}
\title{Add Units}
\usage{
dtt_add_units(x, n = 1L, units = dtt_units(x))

dtt_subtract_units(x, n = 1L, units = dtt_units(x))
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{n}{An integer of the number of units.}

\item{units}{A string of the units.}

\item{...}{Unused.}
}
\value{
The modified Date or POSIXct vector.
}
\description{
Add Units
}
\section{Functions}{
\itemize{
\item \code{dtt_subtract_units}: Subtract Units
}}

\examples{
dtt_add_units(Sys.time())
}
