% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{dtt_add_years}
\alias{dtt_add_years}
\title{Add Years}
\usage{
dtt_add_years(x, n = 1L, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{n}{An integer of the number of years to add}

\item{...}{Unused.}
}
\value{
The modified Date or POSIXct vector.
}
\description{
Add Years
}
\examples{
dtt_add_years(Sys.time())
}
