% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz.R
\name{dtt_adjust_tz}
\alias{dtt_adjust_tz}
\title{Adjust Time Zone}
\usage{
dtt_adjust_tz(x, tz = dtt_sys_tz(), ...)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{tz}{A string of the new time zone.}

\item{...}{Unused.}
}
\value{
The POSIXct vector with the new time zone and time.
}
\description{
Adjust the time zone and time.
}
\examples{
dtt_adjust_tz(Sys.time(), tz = "UTC")
}
\seealso{
\code{\link{dtt_set_tz}}
}
