% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{dtt_aggregate}
\alias{dtt_aggregate}
\title{Aggregate}
\usage{
dtt_aggregate(x, units = dtt_units(x))
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{units}{A string of the units to aggregate by.}
}
\value{
The floored, unique Date or POSIXct vector.
}
\description{
Aggregate
}
\examples{
dtt_aggregate(dates)
}
