% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dtt_complete}
\alias{dtt_complete}
\title{Complete}
\usage{
dtt_complete(x, units, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{units}{A string of the units.}

\item{...}{Unused.}
}
\value{
A flag indicating whether complete.
}
\description{
Completes a Date or POSIXct vector.
}
\examples{
dtt_completed(date_times)
}
