% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dtt_date}
\alias{dtt_date}
\title{Date}
\usage{
dtt_date(x, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
A Date vector.
}
\description{
Date
}
\examples{
dtt_date(Sys.time())
}
