% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-time.R
\name{dtt_date_time}
\alias{dtt_date_time}
\title{Date Time}
\usage{
dtt_date_time(x, tz, ...)
}
\arguments{
\item{x}{A Date vector.}

\item{tz}{A string of the time zone.}

\item{...}{Unused.}
}
\value{
A DateTime vector.
}
\description{
Date Time
}
\examples{
dtt_date_time(Sys.Date())
}
