% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/units.R
\name{dtt_set_day}
\alias{dtt_set_day}
\alias{dtt_day}
\alias{dtt_days}
\alias{dtt_day<-}
\title{Get and Set Days}
\usage{
dtt_set_day(x, value)

dtt_day(x, ...)

dtt_days(x, ...)

dtt_day(x) <- value

dtt_day(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{value}{A numeric vector of the day value(s).}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get and Set Days
}
\examples{
dtt_day(Sys.time())
}
