% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_days}
\alias{dtt_days}
\alias{dtt_day}
\title{Get Days}
\usage{
dtt_days(x, ...)

dtt_day(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Days
}
\examples{
dtt_days(Sys.time())
}
