% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte-time.R, R/dayte.R
\name{dtt_dayte_time.POSIXct}
\alias{dtt_dayte_time.POSIXct}
\alias{dtt_dayte}
\alias{dtt_dayte.Date}
\alias{dtt_dayte.POSIXct}
\title{Dayte}
\usage{
\method{dtt_dayte_time}{POSIXct}(x, start_month = 1L, start_day = 1L,
  ...)

dtt_dayte(x, ...)

\method{dtt_dayte}{Date}(x, start_month = 1L, start_day = 1L, ...)

\method{dtt_dayte}{POSIXct}(x, start_month = 1L, start_day = 1L, ...)
}
\arguments{
\item{x}{A date time vector.}

\item{start_month}{An integer scalar specifying the start month of the year.}

\item{start_day}{An integer scalar specifying the start day of the year.}

\item{...}{Unused.}
}
\value{
A Date vector with the year set to year.

A Date vector of the daytes.
}
\description{
Dayte
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXct}: Dayte Time a POSIXct vector

\item \code{Date}: Dayte a Date vector

\item \code{POSIXct}: Dayte a POSIXct vector
}}

\examples{
dtt_dayte(as.Date(c("2001-01-01", "2015-12-13")))
}
