% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{dtt_daytet}
\alias{dtt_daytet}
\title{Dayte-Time}
\usage{
dtt_daytet(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}
}
\value{
The Date or POSIXct vector with years set to 1972.
}
\description{
Dayte-Time
}
\examples{
dtt_daytet(Sys.Date())
dtt_daytet(Sys.time())
}
