% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{dtt_duration}
\alias{dtt_duration}
\title{Create Duration}
\usage{
dtt_duration(x, units = "seconds")
}
\arguments{
\item{x}{An numeric vector of the duration.}

\item{units}{A string of the time units.}
}
\value{
A dtt_duration object.
}
\description{
Create Duration
}
\examples{
dtt_duration(60)
}
