% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor.R
\name{dtt_floor}
\alias{dtt_floor}
\title{Get Units}
\usage{
dtt_floor(x, units, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{units}{A string of the units to floor by.}

\item{...}{Unused.}
}
\value{
The floored Date or POSIXct vector.
}
\description{
The possible units values are 'seconds', 'minutes', 'hours', 'days', 'months' or 'years'.
}
\examples{
dtt_floor(Sys.time(), "hours")
}
