% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floored.R
\name{dtt_floored}
\alias{dtt_floored}
\title{Floored}
\usage{
dtt_floored(x, units)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{units}{A string of the units to floor by.}

\item{...}{Unused.}
}
\value{
A flag indicating whether floored.
}
\description{
The possible units values are 'seconds', 'minutes', 'hours', 'days', 'months' or 'years'.
}
\examples{
dtt_floored(Sys.time(), "hours")
}
