% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_hour}
\alias{dtt_hour}
\alias{dtt_hours}
\alias{dtt_hours<-}
\alias{dtt_hour<-}
\title{Get and Set Hours}
\usage{
dtt_hour(x, ...)

dtt_hours(x, ...)

dtt_hours(x) <- value

dtt_hour(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}

\item{value}{A numeric vector of the hour value(s).}
}
\value{
An integer vector.
}
\description{
Get and Set Hours
}
\examples{
dtt_hour(Sys.time())
}
