% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_hours}
\alias{dtt_hours}
\alias{dtt_hour}
\title{Get Hours}
\usage{
dtt_hours(x, ...)

dtt_hour(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Hours
}
\examples{
dtt_hours(Sys.time())
}
