% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtt_is_date}
\alias{dtt_is_date}
\title{Is Date}
\usage{
dtt_is_date(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A flag indicating whether R is a Date vector.
}
\description{
Is Date
}
\examples{
dtt_is_date(Sys.Date())
dtt_is_date(Sys.time())
}
