% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/units.R
\name{dtt_set_minute}
\alias{dtt_set_minute}
\alias{dtt_minute}
\alias{dtt_minutes}
\alias{dtt_minute<-}
\title{Get and Set Minutes}
\usage{
dtt_set_minute(x, value)

dtt_minute(x, ...)

dtt_minutes(x, ...)

dtt_minute(x) <- value

dtt_minute(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{value}{A numeric vector of the minute value(s).}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get and Set Minutes
}
\examples{
dtt_minute(Sys.time())
}
