% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_minutes}
\alias{dtt_minutes}
\title{Get Minutes}
\usage{
dtt_minutes(x, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Minutes
}
\examples{
dtt_minutes(Sys.time())
}
