% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/units.R
\name{dtt_set_month}
\alias{dtt_set_month}
\alias{dtt_month}
\alias{dtt_months}
\alias{dtt_months<-}
\alias{dtt_month<-}
\title{Get and Set Months}
\usage{
dtt_set_month(x, value)

dtt_month(x, ...)

dtt_months(x, ...)

dtt_months(x) <- value

dtt_month(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{value}{A numeric vector of the month value(s).}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get and Set Months
}
\examples{
dtt_month(Sys.time())
}
