% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{dtt_rate}
\alias{dtt_rate}
\title{Create Rate}
\usage{
dtt_rate(x, duration = dtt_duration(1L))
}
\arguments{
\item{x}{An double vector of the change.}

\item{duration}{A dtt_duration object of the time interval or 
a string of the units.}
}
\value{
A dtt_duration object.
}
\description{
Create Rate
}
\examples{
dtt_rate(60)
}
