% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/units.R
\name{dtt_set_second}
\alias{dtt_set_second}
\alias{dtt_second}
\alias{dtt_seconds}
\alias{dtt_seconds<-}
\alias{dtt_second<-}
\title{Get and Set Seconds}
\usage{
dtt_set_second(x, value)

dtt_second(x, ...)

dtt_seconds(x, ...)

dtt_seconds(x) <- value

dtt_second(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{value}{A numeric vector of the second value(s).}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get and Set Seconds
}
\examples{
dtt_second(Sys.time())
}
