% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_seconds}
\alias{dtt_seconds}
\title{Get Seconds}
\usage{
dtt_seconds(x, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Seconds
}
\examples{
dtt_seconds(Sys.time())
}
