% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{dtt_seq}
\alias{dtt_seq}
\title{Sequence}
\usage{
dtt_seq(from = dtt_date("2000-01-01"), to = dtt_date("2000-01-01"),
  units = "days", length_out = NULL)
}
\arguments{
\item{from}{A Date or POSIXct scalar of the start date.}

\item{to}{A Date or POSIXct scalar of the to date.}

\item{units}{A string of the units.}

\item{length_out}{An integer of the number of units from from.}
}
\value{
A Date or POSIXct vector.
}
\description{
Creates a unique, ordered complete Date or POSIXct vector.
}
\examples{
dtt_seq(Sys.Date() - 1, Sys.time())
}
