% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz.R
\name{dtt_set_sys_tz}
\alias{dtt_set_sys_tz}
\alias{dtt_reset_sys_tz}
\title{Set System Time Zone}
\usage{
dtt_set_sys_tz(x = NULL)

dtt_reset_sys_tz()
}
\arguments{
\item{x}{A string of the time zone to treat as if its the system time zone
when call \code{\link{dtt_sys_tz}} or \code{NULL} to reset.}
}
\value{
A string of the old value.
}
\description{
Set System Time Zone
}
\section{Functions}{
\itemize{
\item \code{dtt_reset_sys_tz}: Reset System Time Zone
}}

\examples{
\dontrun{
dtt_sys_tz()
old <- dtt_set_sys_tz("Etc/GMT+8")
dtt_sys_tz()
dtt_reset_sys_tz()
dtt_sys_tz()
dtt_set_sys_tz(old)
dtt_sys_tz()
}
}
