% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz.R
\name{dtt_set_tz}
\alias{dtt_set_tz}
\title{Set Time Zone}
\usage{
dtt_set_tz(x, tz = dtt_sys_tz(), ...)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{tz}{A string of the new time zone.}

\item{...}{Unused.}
}
\value{
The POSIXct vector with the new time zone.
}
\description{
Sets the time zone without adjusting the time.
}
\examples{
dtt_set_tz(Sys.time(), tz = "UTC")
}
\seealso{
\code{\link{dtt_adjust_tz}}
}
