% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-year.R
\name{dtt_study_year}
\alias{dtt_study_year}
\title{Study Year}
\usage{
dtt_study_year(x, start_month = 1L, start_day = 1L)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{start_month}{An integer vector specifying the start month of the year.}

\item{start_day}{An integer vector specifying the start day of the year.}
}
\value{
A character vector of the study year.
}
\description{
Study Year
}
\examples{
dtt_study_year(Sys.Date())
dtt_study_year(as.Date(c("2000-03-31", "2000-04-01", "2001-04-01")), start_month = 4L)
}
