% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{dtt_subtract_years}
\alias{dtt_subtract_years}
\alias{dtt_subtract_months}
\alias{dtt_subtract_days}
\alias{dtt_subtract_hours}
\alias{dtt_subtract_minutes}
\alias{dtt_subtract_seconds}
\title{Subtract Years}
\usage{
dtt_subtract_years(x, n = 1L)

dtt_subtract_months(x, n = 1L)

dtt_subtract_days(x, n = 1L)

dtt_subtract_hours(x, n = 1L)

dtt_subtract_minutes(x, n = 1L)

dtt_subtract_seconds(x, n = 1L)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{n}{An integer of the number of years to subtract}

\item{...}{Unused.}
}
\value{
The modified Date or POSIXct vector.
}
\description{
Subtract Years
}
\section{Functions}{
\itemize{
\item \code{dtt_subtract_months}: Subtract Months

\item \code{dtt_subtract_days}: Subtract Days

\item \code{dtt_subtract_hours}: Subtract Hours

\item \code{dtt_subtract_minutes}: Subtract Minutes

\item \code{dtt_subtract_seconds}: Subtract Seconds
}}

\examples{
dtt_subtract_years(Sys.time())
}
