% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz.R
\name{dtt_sys_tz}
\alias{dtt_sys_tz}
\title{Get System Time Zone}
\usage{
dtt_sys_tz()
}
\value{
A string of the time zone.
}
\description{
The user also has the option to set the value returned by \code{dtt_sys_tz()}
using \code{\link{dtt_set_sys_tz}}.
}
\examples{
dtt_sys_tz()
}
