% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{dtt_time}
\alias{dtt_time}
\title{Time}
\usage{
dtt_time(x, ...)
}
\arguments{
\item{x}{The object to coerce.}

\item{...}{Unused.}
}
\value{
A hms vector.
}
\description{
Coerce to a hms object.
}
\examples{
dtt_time(Sys.time())
}
