% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz.R
\name{dtt_tz}
\alias{dtt_tz}
\title{Time Zone}
\usage{
dtt_tz(x, ...)
}
\arguments{
\item{x}{A POSIXct vector.}

\item{...}{Unused.}
}
\value{
A string of the time zone.
}
\description{
Time Zone
}
\examples{
dtt_tz(Sys.time())
}
