% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_units}
\alias{dtt_units}
\title{Get Units}
\usage{
dtt_units(x, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
A string indicating the date time units.
}
\description{
The possible values are 'seconds', 'minutes', 'hours', 'days', 'months' or 'years'.
}
\examples{
dtt_units(Sys.time())
}
