% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/units.R
\name{dtt_set_year}
\alias{dtt_set_year}
\alias{dtt_year}
\alias{dtt_years}
\alias{dtt_years<-}
\alias{dtt_year<-}
\title{Get Years}
\usage{
dtt_set_year(x, value)

dtt_year(x, ...)

dtt_years(x, ...)

dtt_years(x) <- value

dtt_year(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{value}{A numeric vector of the years value(s).}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Years
}
\examples{
dtt_year(Sys.time())
}
