% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_years}
\alias{dtt_years}
\title{Get Years}
\usage{
dtt_years(x, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Years
}
\examples{
dtt_years(Sys.time())
}
