% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{Is POSIXct}
\usage{
is.Date(x)
}
\arguments{
\item{x}{The R object to test.}
}
\value{
A flag indicating whether x inherits from Date.
}
\description{
Tests whether an R object is a Date.
}
