% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date_or_POSIXct}
\alias{is.Date_or_POSIXct}
\title{Is Date or POSIXct}
\usage{
is.Date_or_POSIXct(x)
}
\arguments{
\item{x}{The R object to test.}
}
\value{
A flag indicating whether x inherits from Date or POSIXct.
}
\description{
Tests whether an R object is a Date or POSIXct.
}
