% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{Is POSIXct}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{The R object to test.}
}
\value{
A flag indicating whether x inherits from POSIXct
}
\description{
Tests whether an R object is a POSIXct
}
