% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{times}
\alias{times}
\title{Times}
\format{An object of class \code{hms} (inherits from \code{difftime}) of length 3.}
\usage{
times
}
\description{
A hms vector of length 3.
}
\keyword{datasets}
