% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dtt_complete}
\alias{dtt_complete}
\title{Complete}
\usage{
dtt_complete(x, from = min(x), to = max(x), floor = TRUE,
  unique = TRUE, sort = TRUE, units = dtt_units(x), ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{from}{A date/time scalar of the start.}

\item{to}{A date/time vector of the end.}

\item{floor}{A flag indicating whether to floor the values.}

\item{unique}{A flag indicating whether to return distinct values.}

\item{sort}{A flag indicating whether the returned vector must be sorted.}

\item{units}{A string of the time units.}

\item{...}{Unused.}
}
\value{
The complete and possibly unique and sorted floored vector.
}
\description{
Completes date/time vector.
}
\examples{
dtt_completed(date_times)
}
