% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{dtt_completed}
\alias{dtt_completed}
\title{Completed}
\usage{
dtt_completed(x, unique = TRUE, sorted = TRUE, floor = TRUE,
  units = dtt_units(x), ...)
}
\arguments{
\item{x}{A Date or POSIXct vector}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the values must be sorted.}

\item{floor}{A flag indicating whether to floor the values.}

\item{units}{A string of the units.}

\item{...}{Unused.}
}
\value{
A flag indicating whether complete.
}
\description{
Tests whether a Date or POSIXct object is complete.
}
\examples{
dtt_completed(date_times)
}
