% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dtt_date_time}
\alias{dtt_date_time}
\title{Date}
\usage{
dtt_date_time(x, tz = dtt_sys_tz(), ...)
}
\arguments{
\item{x}{A Date vector.}

\item{tz}{A string of the time zone.}

\item{...}{Unused.}
}
\value{
A DateTime vector.
}
\description{
Date
}
\examples{
dtt_dayte_time(Sys.Date())
}
