% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_days}
\alias{dtt_days}
\alias{dtt_day}
\alias{dtt_days<-}
\alias{dtt_day<-}
\title{Get and Set Days}
\usage{
dtt_days(x, ...)

dtt_day(x, ...)

dtt_days(x) <- value

dtt_day(x) <- value
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}

\item{value}{A numeric vector of the day value(s).}
}
\value{
An integer vector.
}
\description{
Get and Set Days
}
\examples{
dtt_days(Sys.time())
}
