% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dtt_dayte}
\alias{dtt_dayte}
\title{Dayte}
\usage{
dtt_dayte(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
A Date or POSIXct vector with year as 1972.
}
\description{
Dayte
}
\examples{
dtt_dayte(Sys.Date())
}
