% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dtt_dayte_time}
\alias{dtt_dayte_time}
\title{Dayte Time}
\usage{
dtt_dayte_time(x, tz, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{tz}{A string of the time zone.}

\item{...}{Unused.}
}
\value{
A POSIXct vector.
}
\description{
Dayte Time
}
\examples{
dtt_dayte_time(Sys.time())
}
