% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor.R
\name{dtt_floor}
\alias{dtt_floor}
\title{Floor}
\usage{
dtt_floor(x, units, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{units}{A string of the units to floor by.}

\item{...}{Unused.}
}
\value{
The floored date/time vector.
}
\description{
Floors a date/time vector
}
\details{
The possible units values are 'seconds', 'minutes', 'hours', 'days', 'months' or 'years'.
}
\examples{
dtt_floor(Sys.time(), "hours")
}
