% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_months}
\alias{dtt_months}
\alias{dtt_month}
\title{Get Months}
\usage{
dtt_months(x, ...)

dtt_month(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Get Months
}
\examples{
dtt_months(Sys.time())
}
