% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{dtt_seq}
\alias{dtt_seq}
\title{Sequence}
\usage{
dtt_seq(from, to, units = "days")
}
\arguments{
\item{from}{A Date or POSIXct scalar of the start date.}

\item{to}{A Date or POSIXct scalar of the to date.}

\item{units}{A string of the units.}
}
\value{
A Date or POSIXct vector.
}
\description{
Creates a unique, ordered complete Date or POSIXct vector.
}
\examples{
dtt_seq(Sys.Date() - 1, Sys.time())
}
