% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{dtt_time}
\alias{dtt_time}
\title{Time}
\usage{
dtt_time(x, tz, ...)
}
\arguments{
\item{x}{A Date vector.}

\item{tz}{A string of the time zone.}

\item{...}{Unused.}
}
\value{
A hms vector.
}
\description{
Time
}
\examples{
dtt_time(Sys.time())
}
