% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttr-package.R
\docType{package}
\name{dttr-package}
\alias{dttr}
\alias{dttr-package}
\title{dttr: Manipulate Date and Time Objects}
\description{
A light-weight package to manipulate date (Date), time (hms) 
    and datetime (POSIXct) and hms objects.
    As well as functions to get, set, add and subtract years, months, days, hours
    etc it includes functions to floor objects, set or adjust time zones and
    create complete sequences. It also provides duration and timer objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dttr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dttr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
