% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttr-package.R
\docType{package}
\name{dttr-package}
\alias{dttr}
\alias{dttr-package}
\title{dttr: Manipulate Date and POSIXct Vectors}
\description{
Manipulates Date and POSIXct vectors.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
